removed circular dependency by moving stage props out
