+++ src/stageProps.js
@@ -1,0 +1,51 @@
+import PropTypes from "prop-types";
+import { filterByKey, including } from "./utils";
+import { DEFAULT_PROPS, EVENT_PROPS } from "./props";
+
+export function validateCanvas(props, propName, componentName) {
+  // Let's assume that element is canvas if the element is Element and implements getContext
+  const element = props[propName];
+  if (typeof element === "undefined") {
+    return;
+  }
+
+  const isCanvas = element instanceof Element && typeof element.getContext === "function";
+  if (!isCanvas) {
+    const propType = typeof element;
+    return new Error(
+      `Invalid prop '${propName}' of type '${propType}' supplied to '${componentName}', expected '<canvas> Element'.`
+    );
+  }
+}
+
+export const propTypes = {
+  options: PropTypes.shape({
+    antialias: PropTypes.bool,
+    autoStart: PropTypes.bool,
+    backgroundColor: PropTypes.number,
+    clearBeforeRender: PropTypes.bool,
+    forceCanvas: PropTypes.bool,
+    forceFXAA: PropTypes.bool,
+    height: PropTypes.number,
+    legacy: PropTypes.bool,
+    powerPreference: PropTypes.string,
+    preserveDrawingBuffer: PropTypes.bool,
+    resolution: PropTypes.number,
+    roundPixels: PropTypes.bool,
+    sharedLoader: PropTypes.bool,
+    sharedTicker: PropTypes.bool,
+    transparent: PropTypes.bool,
+    view: validateCanvas,
+    width: PropTypes.number,
+  }),
+  children: PropTypes.node,
+  height: PropTypes.number,
+  width: PropTypes.number,
+};
+
+export const includingDisplayObjectProps = including(Object.keys(DEFAULT_PROPS).concat(EVENT_PROPS));
+export const includingStageProps = including(Object.keys(propTypes));
+export const includingCanvasProps = key => !includingDisplayObjectProps(key) && !includingStageProps(key);
+
+export const getCanvasProps = props => filterByKey(props, includingCanvasProps);
+export const getDisplayObjectProps = props => filterByKey(props, includingDisplayObjectProps);
